refactor: move functionalities back to generic types.
