+++ src/Ursa/Controls/DateTimePicker/Base/DateTimePickerBaseT.cs
@@ -2,12 +2,16 @@ using Avalonia;
 using Avalonia.Controls;
 using Avalonia.Controls.Primitives;
 using Avalonia.Data;
+using Avalonia.Input;
+using Avalonia.Interactivity;
+using Avalonia.Media;
 using Irihi.Avalonia.Shared.Common;
 using Irihi.Avalonia.Shared.Contracts;
+using Irihi.Avalonia.Shared.Helpers;
 
 namespace Ursa.Controls;
 
-public abstract class DateTimePickerBase<T> : DateTimePickerBase where T : struct
+public abstract class DateTimePickerBase<T> : DateTimePickerBase, IClearControl where T : struct
 {
     public static readonly StyledProperty<T?> SelectedDateProperty =
         AvaloniaProperty.Register<DateTimePickerBase<T>, T?>(
@@ -18,6 +22,11 @@ public abstract class DateTimePickerBase<T> : DateTimePickerBase where T : struc
         get => GetValue(SelectedDateProperty);
         set => SetValue(SelectedDateProperty, value);
     }
+
+    protected TextBox? _textBox;
+    protected DatePickerCalendarView? _calendar;
+    protected Popup? _popup;
+    protected TimePickerPresenter? _timePickerPresenter;
 
     /// <summary>Extracts the date component from a <typeparamref name="T"/> value.</summary>
     protected abstract DateOnly? ToDateOnly(T? value);
@@ -40,15 +49,66 @@ public abstract class DateTimePickerBase<T> : DateTimePickerBase where T : struc
     /// <summary>Returns the current time for use when no time component is available.</summary>
     protected virtual TimeOnly GetCurrentTime() => TimeOnly.FromDateTime(DateTime.Now);
 
-    protected override DateOnly? GetSelectedDateOnly() => ToDateOnly(SelectedDate);
+    protected override void OnApplyTemplate(TemplateAppliedEventArgs e)
+    {
+        base.OnApplyTemplate(e);
 
-    protected override TimeOnly? GetSelectedTimeOnly() =>
+        TimePickerPresenter.SelectedTimeChangedEvent.RemoveHandler(OnTimeSelected, _timePickerPresenter);
+        DatePickerCalendarView.DateSelectedEvent.RemoveHandler(OnDateSelected, _calendar);
+        GotFocusEvent.RemoveHandler(OnTextBoxGotFocus, _textBox);
+        PointerPressedEvent.RemoveHandler(OnTextBoxPressed, _textBox);
+        LostFocusEvent.RemoveHandler(OnTextBoxLostFocus, _textBox);
+
+        _popup = e.NameScope.Find<Popup>(PART_Popup);
+        _textBox = e.NameScope.Find<TextBox>(PART_TextBox);
+        _calendar = e.NameScope.Find<DatePickerCalendarView>(PART_Calendar);
+        _timePickerPresenter = e.NameScope.Find<TimePickerPresenter>(PART_TimePicker);
+
+        TimePickerPresenter.SelectedTimeChangedEvent.AddHandler(OnTimeSelected, _timePickerPresenter);
+        DatePickerCalendarView.DateSelectedEvent.AddHandler(OnDateSelected, _calendar);
+        GotFocusEvent.AddHandler(OnTextBoxGotFocus, _textBox);
+        PointerPressedEvent.AddHandler(OnTextBoxPressed, RoutingStrategies.Tunnel, true, _textBox);
+        LostFocusEvent.AddHandler(OnTextBoxLostFocus, _textBox);
+
+        SyncToUI();
+    }
+
+    private void OnTextBoxLostFocus(object? sender, FocusChangedEventArgs e) => CommitInput();
+    private void OnTextBoxGotFocus(object? sender, FocusChangedEventArgs e) => InitializePopupOpen();
+    private void OnTextBoxPressed(object? sender, PointerPressedEventArgs e) => InitializePopupOpen();
+
+    private void InitializePopupOpen()
+    {
+        SetCurrentValue(IsDropdownOpenProperty, true);
+        var date = GetCalendarContextDate();
+        _calendar?.SyncContextDate(new DatePickerCalendarContext(date.Year, date.Month));
+        _timePickerPresenter?.SyncTime(GetSelectedTimeOnly());
+    }
+
+    private void OnDateSelected(object? sender, DatePickerCalendarDayButtonEventArgs e)
+    {
+        if (e.Date is null) return;
+        var time = SelectedDate.HasValue
+            ? ToTimeOnly(SelectedDate.Value) ?? GetCurrentTime()
+            : GetCurrentTime();
+        SetCurrentValue(SelectedDateProperty, (T?)CombineDateTime(e.Date.Value, time));
+    }
+
+    private void OnTimeSelected(object? sender, TimeChangedEventArgs e)
+    {
+        if (e.NewTime is null) return;
+        var date = GetSelectedDateOnly() ?? GetToday();
+        SetCurrentValue(SelectedDateProperty, (T?)CombineDateTime(date, e.NewTime.Value));
+    }
+
+    private DateOnly? GetSelectedDateOnly() => ToDateOnly(SelectedDate);
+
+    private TimeOnly? GetSelectedTimeOnly() =>
         SelectedDate.HasValue ? ToTimeOnly(SelectedDate.Value) : null;
 
-    protected override DateOnly GetCalendarContextDate() =>
-        GetSelectedDateOnly() ?? GetToday();
+    private DateOnly GetCalendarContextDate() => GetSelectedDateOnly() ?? GetToday();
 
-    protected override void SyncToUI()
+    private void SyncToUI()
     {
         if (SelectedDate is null)
         {
@@ -66,7 +126,7 @@ public abstract class DateTimePickerBase<T> : DateTimePickerBase where T : struc
         }
     }
 
-    protected override void CommitInput()
+    private void CommitInput()
     {
         var format = DisplayFormat ?? DEFAULT_DATETIME_DISPLAY_FORMAT;
         if (string.IsNullOrWhiteSpace(_textBox?.Text))
@@ -78,18 +138,29 @@ public abstract class DateTimePickerBase<T> : DateTimePickerBase where T : struc
         SetCurrentValue(SelectedDateProperty, parsed);
     }
 
-    protected override void OnCalendarDateSelected(DateOnly date)
+    protected override void OnPointerPressed(PointerPressedEventArgs e)
     {
-        var time = SelectedDate.HasValue
-            ? ToTimeOnly(SelectedDate.Value) ?? GetCurrentTime()
-            : GetCurrentTime();
-        SetCurrentValue(SelectedDateProperty, (T?)CombineDateTime(date, time));
+        base.OnPointerPressed(e);
+        if (!e.Handled && e.Source is Visual source)
+        {
+            if (_popup?.IsInsidePopup(source) == true)
+                e.Handled = true;
+            else
+                InitializePopupOpen();
+        }
     }
 
-    protected override void OnTimePickerTimeSelected(TimeOnly time)
+    protected override void OnLostFocus(FocusChangedEventArgs e)
     {
-        var date = GetSelectedDateOnly() ?? GetToday();
-        SetCurrentValue(SelectedDateProperty, (T?)CombineDateTime(date, time));
+        base.OnLostFocus(e);
+        var newItem = e.NewFocusedElement;
+        if (Equals(newItem, _textBox)) return;
+        if (newItem is Visual visual)
+        {
+            if (_popup?.IsInsidePopup(visual) == true) return;
+            CommitInput();
+            SetCurrentValue(IsDropdownOpenProperty, false);
+        }
     }
 
     protected override void OnPropertyChanged(AvaloniaPropertyChangedEventArgs change)
@@ -102,5 +173,5 @@ public abstract class DateTimePickerBase<T> : DateTimePickerBase where T : struc
         }
     }
 
-    public override void Clear() => SetCurrentValue(SelectedDateProperty, (T?)null);
+    public void Clear() => SetCurrentValue(SelectedDateProperty, (T?)null);
 }
